requirements/dev.txt
-r base.txt

# pytest requirements
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
toml==0.10.2

# Debugging requirements
django-debug-toolbar==3.2.4

# Faker data
Faker==13.13.0

# Sentry
sentry-sdk-1.9.3